requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
sqlparse==0.4.4
typing-extensions==4.8.0
djangorestframework==3.14.0